Some draft control files for the low band
Draft fourfit control files that contain flagging information for
tracks e17d05, e17a10, and e17e11.  The flag data are based on both
the information contained in the scan logs and dropouts seen in
segmented data post-fourfit.  This information could be folded into
the flag database solution that will be implemented.

Manual phase cals have been calculated for each day using `phases.pl`,
so we can fourfit the data and end up with reasonably flat phases as a
function of freq.  However, polarization delay has not been adjusted
and sbd and mbd may be different (modulo mbd ambiguities).

A good first step for the HOPS fringe-fitting group would be to nail
down manual phase cals and delay offsets for each day so that residual
delays are as described above.  Once this is done, tight sbd windows
can be set on non-detections.  A good second step (although it
factorizes with the first) is to derive adhoc phase calibration
solutions.  Both of these can probably be done for most sources/scans
using data from ALMA baselines alone, without the need to globalize.